Handle entire Webhook payload generically
Previously, the implementation of this resource special-cased a small number
of the Slack Webhook options that were available.  Specifically, this excluded
the use of attachments to create richer messages.  This change re-implements
the out script to allow any properly structured Slack payload.  It also
ensures that environment variable interpolation occurs in any string in that
payload.  One of the best side-effects is that the implementation is nearly
60% smaller and is much less complex.

This change also updates the Docker image to be based on alpine, resulting in
a drastically smaller image with the same functionality.