Concourse Pipeline for Versioned Docker Images
New ci/ directory houses pipeline configuration, release scripts and a
repipe script for reconfiguring a target concourse installation.

To use:

    ./ci/repipe

Changes to the configuration of the pipeline should be done in
ci/settings.yml, which will be spruce-merged over the top of
ci/pipeline.yml.  Credentials are stored in S&W vault.

The git repo will be tagged with a new version everytime the `promote` job
is manually triggered, to help people diagnose bugs with specific versions
of the Docker image.

Dockerhub will contain the following image tags:

  - **latest** - The latest promoted build of this resource image
  - **edge** - The latest build from master that has passed testing
  - **vX.Y.Z** - The image at that specific version