Interpolate the 'text' field in attachments. (#27)
Fixes #25.

The gist of this change is to iterate over the attachments, pulling out
the `text' field and eval/printfing it, and then merging them back
together using jq's `*' operator.  This may not be the right way to do
this.  This may not be the best way to do this, but it does seem to
work, and that's got to count for something.