Added how to mention users to the README.md
Added instructions for how to mention users in the text param to the `README.md`, along with a warning that the old format is no longer supported. Provided links to the official changelogs and the documentation where users can read more regarding this change and how to format their messages in general.